update copyright header
