--- platform/lib/samples/examples/src/responsive.html
@@ -1,55 +1,0 @@
-<!doctype html>
-<html ⚡>
-<head>
-  <meta charset="utf-8">
-  <script async src="https://cdn.ampproject.org/v0.js"></script>
-  <script async custom-element="amp-youtube" src="https://cdn.ampproject.org/v0/amp-youtube-0.1.js"></script>
-  <meta name="viewport" content="width=device-width,minimum-scale=1,initial-scale=1">
-  <style amp-boilerplate>body{-webkit-animation:-amp-start 8s steps(1,end) 0s 1 normal both;-moz-animation:-amp-start 8s steps(1,end) 0s 1 normal both;-ms-animation:-amp-start 8s steps(1,end) 0s 1 normal both;animation:-amp-start 8s steps(1,end) 0s 1 normal both}@-webkit-keyframes -amp-start{from{visibility:hidden}to{visibility:visible}}@-moz-keyframes -amp-start{from{visibility:hidden}to{visibility:visible}}@-ms-keyframes -amp-start{from{visibility:hidden}to{visibility:visible}}@-o-keyframes -amp-start{from{visibility:hidden}to{visibility:visible}}@keyframes -amp-start{from{visibility:hidden}to{visibility:visible}}</style><noscript><style amp-boilerplate>body{-webkit-animation:none;-moz-animation:none;-ms-animation:none;animation:none}</style></noscript>
-</head>
-<body>
-  <!-- Youtube -->
-  <amp-youtube data-videoid="WrpkFROqR0Q"
-               layout="responsive"
-               width="560" height="315">
-  </amp-youtube>
-  <!-- resolution  -->
-  <amp-img alt="apple" src="apple.jpg"
-      height="596" width="900"
-      srcset="apple-900.jpg 900w,
-              apple-800.jpg 800w,
-              apple-700.jpg 700w,
-              apple-600.jpg 600w,
-              apple-500.jpg 500w,
-              apple-400.jpg 400w"
-       sizes="(max-width: 400px) 100vw, 
-              (max-width: 900px) 75vw, 600px">
-  </amp-img>
-  <!-- breakpoints  -->
-  <div>
-    <amp-img alt="grey cat"
-      media="(min-width: 670px)"
-      width="650" height="340"
-      src="cat-large.jpg"></amp-img>
-    <amp-img alt="grey cat"
-      media="(min-width: 470px) and (max-width: 669px)"
-      width="450" height="340"
-      src="cat-medium.jpg"></amp-img>
-    <amp-img alt="grey cat"
-      media="(max-width: 469px)"
-      width="226" height="340"
-      src="cat-small.jpg"></amp-img>
-  </div>
-  <!-- webp -->
-  <amp-img alt="Mountains"
-    width="550" height="368"
-    layout="responsive"
-    src="mountains.webp">
-  <amp-img alt="Mountains"
-      fallback
-      width="550" height="368"
-      layout="responsive"
-      src="mountains.jpg"></amp-img>
-  </amp-img>
-</body>
-</html>
